requirements.txt
torch
torchvision
transformers
pillow
PyMuPDF
scikit-learn
numpy 